NP-2170 Restructure files
