+++ src/pages/public_registration/PublicFilesContent.tsx
@@ -13,7 +13,7 @@ import { setNotification } from '../../redux/actions/notificationActions';
 import { NotificationVariant } from '../../types/notification.types';
 import ButtonWithProgress from '../../components/ButtonWithProgress';
 import { PublicRegistrationContentProps } from './PublicRegistrationContent';
-import { PreviewFile } from './PreviewFile';
+import PreviewFile from './preview_file';
 
 const StyledFileRowContainer = styled.div`
   > :not(:last-child) {
