+++ src/pages/public_registration/preview_file/PreviewFile.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import { File } from '../../../types/file.types';
+import PreviewImg from './PreviewImg';
+import PreviewPdf from './PreviewPdf';
+
+export interface CommonPreviewProps {
+  url: string;
+}
+
+export interface PreviewFileProps extends CommonPreviewProps {
+  file: File;
+}
+
+const PreviewFile = ({ url, file, ...props }: PreviewFileProps) => {
+  const fileType = file.mimeType.toLowerCase();
+
+  return fileType.includes('pdf') ? (
+    <PreviewPdf url={url} {...props} />
+  ) : fileType.includes('image') ? (
+    <PreviewImg url={url} file={file} {...props} />
+  ) : null;
+};
+
+export default PreviewFile;
