+++ src/pages/public_registration/preview_file/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PreviewFile';
